* Relocated fakes * Updated README
